refactor(rust): move put_at methods from buffer.rs to unsafe_util.rs
