+++ rust/fory-core/src/buffer.rs
@@ -514,233 +514,6 @@ impl<'a> Writer<'a> {
         debug_assert!(new_len <= self.bf.capacity());
         self.bf.set_len(new_len);
     }
-
-    /// # Safety
-    #[inline(always)]
-    pub unsafe fn put_bool_at(ptr: *mut u8, value: bool) -> usize {
-        *ptr = value as u8;
-        1
-    }
-
-    /// # Safety
-    #[inline(always)]
-    pub unsafe fn put_i8_at(ptr: *mut u8, value: i8) -> usize {
-        *ptr = value as u8;
-        1
-    }
-
-    /// # Safety
-    #[inline(always)]
-    pub unsafe fn put_u8_at(ptr: *mut u8, value: u8) -> usize {
-        *ptr = value;
-        1
-    }
-
-    /// # Safety
-    #[inline(always)]
-    pub unsafe fn put_i16_at(ptr: *mut u8, value: i16) -> usize {
-        Self::put_u16_at(ptr, value as u16)
-    }
-
-    /// # Safety
-    #[inline(always)]
-    pub unsafe fn put_u16_at(ptr: *mut u8, value: u16) -> usize {
-        std::ptr::copy_nonoverlapping(&value.to_le() as *const u16 as *const u8, ptr, 2);
-        2
-    }
-
-    /// # Safety
-    #[inline(always)]
-    pub unsafe fn put_i32_at(ptr: *mut u8, value: i32) -> usize {
-        Self::put_u32_at(ptr, value as u32)
-    }
-
-    /// # Safety
-    #[inline(always)]
-    pub unsafe fn put_u32_at(ptr: *mut u8, value: u32) -> usize {
-        std::ptr::copy_nonoverlapping(&value.to_le() as *const u32 as *const u8, ptr, 4);
-        4
-    }
-
-    /// # Safety
-    #[inline(always)]
-    pub unsafe fn put_i64_at(ptr: *mut u8, value: i64) -> usize {
-        Self::put_u64_at(ptr, value as u64)
-    }
-
-    /// # Safety
-    #[inline(always)]
-    pub unsafe fn put_u64_at(ptr: *mut u8, value: u64) -> usize {
-        std::ptr::copy_nonoverlapping(&value.to_le() as *const u64 as *const u8, ptr, 8);
-        8
-    }
-
-    /// # Safety
-    #[inline(always)]
-    pub unsafe fn put_f16_at(ptr: *mut u8, value: float16) -> usize {
-        Self::put_u16_at(ptr, value.to_bits())
-    }
-
-    /// # Safety
-    #[inline(always)]
-    pub unsafe fn put_f32_at(ptr: *mut u8, value: f32) -> usize {
-        std::ptr::copy_nonoverlapping(&value as *const f32 as *const u8, ptr, 4);
-        4
-    }
-
-    /// # Safety
-    #[inline(always)]
-    pub unsafe fn put_f64_at(ptr: *mut u8, value: f64) -> usize {
-        std::ptr::copy_nonoverlapping(&value as *const f64 as *const u8, ptr, 8);
-        8
-    }
-
-    /// # Safety
-    #[inline(always)]
-    pub unsafe fn put_i128_at(ptr: *mut u8, value: i128) -> usize {
-        Self::put_u128_at(ptr, value as u128)
-    }
-
-    /// # Safety
-    #[inline(always)]
-    pub unsafe fn put_u128_at(ptr: *mut u8, value: u128) -> usize {
-        std::ptr::copy_nonoverlapping(&value.to_le() as *const u128 as *const u8, ptr, 16);
-        16
-    }
-
-    /// # Safety
-    #[inline(always)]
-    pub unsafe fn put_varint32_at(ptr: *mut u8, value: i32) -> usize {
-        let zigzag = ((value as i64) << 1) ^ ((value as i64) >> 31);
-        Self::put_var_uint32_at(ptr, zigzag as u32)
-    }
-
-    /// # Safety
-    #[inline(always)]
-    pub unsafe fn put_var_uint32_at(ptr: *mut u8, value: u32) -> usize {
-        if value < 0x80 {
-            *ptr = value as u8;
-            1
-        } else if value < 0x4000 {
-            *ptr = ((value as u8) & 0x7F) | 0x80;
-            *ptr.add(1) = (value >> 7) as u8;
-            2
-        } else if value < 0x200000 {
-            *ptr = ((value as u8) & 0x7F) | 0x80;
-            *ptr.add(1) = (((value >> 7) as u8) & 0x7F) | 0x80;
-            *ptr.add(2) = (value >> 14) as u8;
-            3
-        } else if value < 0x10000000 {
-            *ptr = ((value as u8) & 0x7F) | 0x80;
-            *ptr.add(1) = (((value >> 7) as u8) & 0x7F) | 0x80;
-            *ptr.add(2) = (((value >> 14) as u8) & 0x7F) | 0x80;
-            *ptr.add(3) = (value >> 21) as u8;
-            4
-        } else {
-            *ptr = ((value as u8) & 0x7F) | 0x80;
-            *ptr.add(1) = (((value >> 7) as u8) & 0x7F) | 0x80;
-            *ptr.add(2) = (((value >> 14) as u8) & 0x7F) | 0x80;
-            *ptr.add(3) = (((value >> 21) as u8) & 0x7F) | 0x80;
-            *ptr.add(4) = (value >> 28) as u8;
-            5
-        }
-    }
-
-    /// # Safety
-    #[inline(always)]
-    pub unsafe fn put_varint64_at(ptr: *mut u8, value: i64) -> usize {
-        let zigzag = ((value << 1) ^ (value >> 63)) as u64;
-        Self::put_var_uint64_at(ptr, zigzag)
-    }
-
-    /// # Safety
-    #[inline(always)]
-    pub unsafe fn put_var_uint64_at(ptr: *mut u8, value: u64) -> usize {
-        if value < 0x80 {
-            *ptr = value as u8;
-            return 1;
-        }
-        *ptr = ((value as u8) & 0x7F) | 0x80;
-        if value < 0x4000 {
-            *ptr.add(1) = (value >> 7) as u8;
-            return 2;
-        }
-        *ptr.add(1) = (((value >> 7) as u8) & 0x7F) | 0x80;
-        if value < 0x200000 {
-            *ptr.add(2) = (value >> 14) as u8;
-            return 3;
-        }
-        *ptr.add(2) = (((value >> 14) as u8) & 0x7F) | 0x80;
-        if value < 0x10000000 {
-            *ptr.add(3) = (value >> 21) as u8;
-            return 4;
-        }
-        *ptr.add(3) = (((value >> 21) as u8) & 0x7F) | 0x80;
-        if value < 0x800000000 {
-            *ptr.add(4) = (value >> 28) as u8;
-            return 5;
-        }
-        *ptr.add(4) = (((value >> 28) as u8) & 0x7F) | 0x80;
-        if value < 0x40000000000 {
-            *ptr.add(5) = (value >> 35) as u8;
-            return 6;
-        }
-        *ptr.add(5) = (((value >> 35) as u8) & 0x7F) | 0x80;
-        if value < 0x2000000000000 {
-            *ptr.add(6) = (value >> 42) as u8;
-            return 7;
-        }
-        *ptr.add(6) = (((value >> 42) as u8) & 0x7F) | 0x80;
-        if value < 0x100000000000000 {
-            *ptr.add(7) = (value >> 49) as u8;
-            return 8;
-        }
-        *ptr.add(7) = (((value >> 49) as u8) & 0x7F) | 0x80;
-        *ptr.add(8) = (value >> 56) as u8;
-        9
-    }
-
-    /// # Safety
-    #[inline(always)]
-    pub unsafe fn put_tagged_i64_at(ptr: *mut u8, value: i64) -> usize {
-        const HALF_MIN: i64 = i32::MIN as i64 / 2;
-        const HALF_MAX: i64 = i32::MAX as i64 / 2;
-        if (HALF_MIN..=HALF_MAX).contains(&value) {
-            let v = (value as i32) << 1;
-            std::ptr::copy_nonoverlapping(&v.to_le() as *const i32 as *const u8, ptr, 4);
-            4
-        } else {
-            *ptr = 0b1;
-            std::ptr::copy_nonoverlapping(&value.to_le() as *const i64 as *const u8, ptr.add(1), 8);
-            9
-        }
-    }
-
-    /// # Safety
-    #[inline(always)]
-    pub unsafe fn put_tagged_u64_at(ptr: *mut u8, value: u64) -> usize {
-        if value <= i32::MAX as u64 {
-            let v = (value as u32) << 1;
-            std::ptr::copy_nonoverlapping(&v.to_le() as *const u32 as *const u8, ptr, 4);
-            4
-        } else {
-            *ptr = 0b1;
-            std::ptr::copy_nonoverlapping(&value.to_le() as *const u64 as *const u8, ptr.add(1), 8);
-            9
-        }
-    }
-
-    /// # Safety
-    #[inline(always)]
-    pub unsafe fn put_usize_at(ptr: *mut u8, value: usize) -> usize {
-        const SIZE: usize = std::mem::size_of::<usize>();
-        match SIZE {
-            2 => Self::put_u16_at(ptr, value as u16),
-            4 => Self::put_var_uint32_at(ptr, value as u32),
-            8 => Self::put_var_uint64_at(ptr, value as u64),
-            _ => unreachable!(),
-        }
-    }
 }
 
 #[derive(Default)]
